New hitbox modules and registry
