+++ arcade/hitbox/base.py
@@ -1,0 +1,15 @@
+from typing import Tuple
+from PIL.Image import Image
+from arcade import PointList
+
+
+class HitBoxAlgorithm:
+    """
+    Base class for hit box algorithms. Hit box algorithms are used to calculate the
+    points that make up a hit box for a sprite.
+    """
+    name = "base"
+
+    @classmethod
+    def calculate(cls, image: Image, **kwargs) -> PointList:
+        raise NotImplementedError
